flask-osiris/requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_session
psycopg2
cryptography